상세조회 기능, 모든 데이터 출력이 아닌 제목 작성자명 작성날짜 작성내용만 출력하도록 수정, 패스워드 체크해서 맞으면 수정과 삭제가 기능하도록 수정 삭제기능 반환타입 String형으로 삭제 완료와 암호가 틀립니다. 출력.

+++ src/main/java/com/sparta/blog/service/PostService.java
@@ -1,6 +1,7 @@
 package com.sparta.blog.service;
 
 import com.sparta.blog.Entity.Post;
+import com.sparta.blog.dto.PostDetailDto;
 import com.sparta.blog.dto.PostRequestDto;
 import com.sparta.blog.dto.PostResponseDto;
 import com.sparta.blog.repository.PostRepository;
@@ -31,24 +32,39 @@ public class PostService {
        return postResponseDto;
     }
 
-    public List<PostResponseDto> getPost() {
+    public List<PostDetailDto> getPost() {
         //DB조회하기
-        return postRepository.findAll().stream().map(PostResponseDto::new).toList();
+        return postRepository.findAllByOrderByCreateAtDesc().stream().map(PostDetailDto::new).toList();
     }
     @Transactional
-    public Long updatePost(Long id, PostRequestDto requestDto){
+    public PostRequestDto updatePost(Long id, PostRequestDto requestDto, String password){
+
         Post post = findPost(id);
-        post.update(requestDto);
+        if(password.equals(post.getPassword())) {
+            post.update(requestDto);
 
-        return id;
+            return requestDto;
+        }
+        else{
+            throw new IllegalArgumentException("암호가 틀립니다.");
+        }
     }
-    public Long deletePost(Long id){
+    public String deletePost(Long id,String password){
+
         Post post = findPost(id);
-
-        postRepository.delete(post);
-        return id;
+        if(password.equals(post.getPassword())) {
+            postRepository.delete(post);
+            return "삭제완료";
+        }
+        else{
+            return "암호가 틀렸습니다.";
+        }
     }
-
+    public PostDetailDto detailInquiry(Long id) {
+        Post post = findPost(id);
+        PostDetailDto postDetailDto = new PostDetailDto(post);
+        return postDetailDto;
+    }
     private Post findPost(Long id){
         Post post = postRepository.findById(id).orElseThrow(() ->
                 new IllegalArgumentException("선택한 게시글이 존재하지 않습니다.")
@@ -56,4 +72,6 @@ public class PostService {
         );
         return post;
     }
+
+
 }
